view,updateProjects

+++ FundMeUp/Models/BackerProject.cs
@@ -10,7 +10,7 @@ namespace FundMeUp.Models
     public Project Project { get; set; }
     public int ProjectId { get; set; }
     public DateTime DoF { get; set; }
-    public decimal Fund { get; set; }
+    public float Fund { get; set; }
     public Reward Reward { get; set;}
     public int RewardId { get; set; }
   }
